Refactor HaXe and Lua
Now both codebases are as similar as possible.
Changing one should be trivial with the other.
HaXe has a few problems with cast and utils.Draw
(both do not work on Windows target) and with
its html5 target (too slow). So I'm going to
continue using both cast and utils.Draw, as I
will only need to target flash.